starting implementing stdin commands

+++ pj2/heap.h
@@ -4,7 +4,7 @@
 void min_heapify(HEAP *heap, int elementIndex);
 void build_min_heap(HEAP *heap);
 void heapsort(HEAP *heap);
-int heap_min(HEAP *heap);
-int heap_extract_min(HEAP *heap);
-void heap_decrease_key(HEAP *heap, int elem, int key);
-void min_heap_insert(HEAP *heap, int key);
+double heap_min(HEAP *heap);
+double heap_extract_min(HEAP *heap);
+void heap_decrease_key(HEAP *heap, int elem, double key);
+void min_heap_insert(HEAP *heap, double key);
